study(step2): Player 추가
