+++ src/main/kotlin/blackjack/domain/Gamer.kt
@@ -1,29 +1,3 @@
 package blackjack.domain
 
-data class Gamer(val name: String, val cards: Cards = Cards()) {
-
-    override fun toString(): String {
-        return "${name}카드: $cards"
-    }
-
-    fun getTotalScore(): Int {
-        val totalScoreByAceEleven = cards.getTotalScore(isAceEleven = true)
-        val diffByAceEleven = totalScoreByAceEleven - BLACK_JACK
-        val totalScoreByAceOne = cards.getTotalScore(isAceEleven = false)
-        val diffByAceOne = totalScoreByAceOne - BLACK_JACK
-
-        return if (diffByAceEleven < diffByAceOne) {
-            totalScoreByAceEleven
-        } else {
-            totalScoreByAceOne
-        }
-    }
-
-    fun addCard(card: Card) = cards.addCard(card)
-
-    fun hasCard(card: Card): Boolean = cards.hasCard(card)
-
-    companion object {
-        private const val BLACK_JACK = 21
-    }
-}
+data class Gamer(override val name: String, override val cards: Cards = Cards()) : Player(name, cards)
